Cumulative risk of ovarian cancer in women with the BRCA1 mutation.

BRCA1_ova provides the cumulative risk of ovarian cancer in a population of women with the BRCA1 mutation as a function of their age (in years).

Usage

BRCA1_ova

Format

A data frame (63 x 2).

age: age (in years).

cumRisk: cumulative risk of developing ovarian cancer in this (BRCA1) population.

Source

Based on Figure 2 (p. 2408) of Kuchenbaecker, K. B., Hopper, J. L., Barnes, D. R., Phillips, K. A., Mooij, T. M., Roos-Blom, M. J., ... & BRCA1 and BRCA2 Cohort Consortium (2017). Risks of breast, ovarian, and contralateral breast cancer for BRCA1 and BRCA2 mutation carriers. JAMA, 317 (23), 2402–2416. doi: 10.1001/jama.2017.7112

See also