Cumulative risk of breast cancer in women with the BRCA2 mutation.

BRCA2_mam provides the cumulative risk of breast cancer in a population of women with the BRCA2 mutation as a function of their age (in years).

BRCA2_mam

Format

A data frame (63 x 2).

age: age (in years).

cumRisk: cumulative risk of developing breast cancer in this (BRCA2) population.

Source

Based on Figure 2 (p. 2408) of Kuchenbaecker, K. B., Hopper, J. L., Barnes, D. R., Phillips, K. A., Mooij, T. M., Roos-Blom, M. J., ... & BRCA1 and BRCA2 Cohort Consortium (2017). Risks of breast, ovarian, and contralateral breast cancer for BRCA1 and BRCA2 mutation carriers. JAMA, 317 (23), 2402--2416. doi: 10.1001/jama.2017.7112

See also