Convert from FFTrees to riskyr objects.

FFTrees_riskyr converts an FFTrees object --- as generated by the FFTrees package --- into a corresponding riskyr object.

FFTrees_riskyr(x, data = "train", tree = 1)

Arguments

x

An FFTrees object (generated by FFTrees).

data

The type of data to consider (as a character string). Must be either "train" (for training/fitting data) or "test" (for test/prediction data). Default: data = "train".

tree

An integer specifying the tree to consider (as an integer). Default: tree = 1.

Value

A riskyr scenario (as riskyr object).

Details

FFTrees_riskyr essentially allows using riskyr functions to visualize a fast-and-frugal tree (FFT)'s performance information (as contained in a 2x2 matrix of frequency counts).

The R package FFTrees creates, visualizes, and evaluates fast-and-frugal trees (FFTs) for solving binary classification problems in an efficient and transparent fashion.

References

See https://CRAN.R-project.org/package=FFTrees or https://github.com/ndphillips/FFTrees for information on the R package FFTrees.

See also