The positive predictive value of a decision process or diagnostic procedure.

PPV defines some decision's positive predictive value (PPV): The conditional probability of the condition being TRUE provided that the decision is positive.

Usage

PPV

Format

An object of class numeric of length 1.

Details

Understanding or obtaining the positive predictive value PPV:

Definition: PPV is the conditional probability for the condition being TRUE given a positive decision:

PPV = p(condition = TRUE | decision = positive)

or the probability of a positive decision being correct.

Perspective: PPV further classifies the subset of dec_pos individuals by condition (PPV = hi/dec_pos = hi/(hi + fa)).

Alternative names: precision

Relationships:

a. PPV is the complement of the false discovery or false detection rate FDR:

PPV = 1 - FDR

b. PPV is the opposite conditional probability – but not the complement – of the sensitivity sens:

sens = p(decision = positive | condition = TRUE)

In terms of frequencies, PPV is the ratio of hi divided by dec_pos (i.e., hi + fa):

PPV = hi/dec_pos = hi/(hi + fa)

Dependencies: PPV is a feature of a decision process or diagnostic procedure and – similar to the sensitivity sens – a measure of correct decisions (positive decisions that are actually TRUE).

However, due to being a conditional probability, the value of PPV is not intrinsic to the decision process, but also depends on the condition's prevalence value prev.

References

Consult Wikipedia for additional information.

See also

Examples

PPV <- .55     # sets a positive predictive value of 55%
PPV <- 55/100  # (condition = TRUE) for 55 out of 100 people with (decision = positive)
is_prob(PPV)   # TRUE
#> [1] TRUE