Compute a decision's false omission rate (FOR) from probabilities.

comp_FOR computes the false omission rate FOR from 3 essential probabilities prev, sens, and spec.

comp_FOR(prev, sens, spec)

Arguments

prev

The condition's prevalence prev (i.e., the probability of condition being TRUE).

sens

The decision's sensitivity sens (i.e., the conditional probability of a positive decision provided that the condition is TRUE).

spec

The decision's specificity value spec (i.e., the conditional probability of a negative decision provided that the condition is FALSE).

Value

The false omission rate FOR as a probability. A warning is provided for NaN values.

Details

comp_FOR uses probabilities (not frequencies) and does not round results.

See also

Examples

# (1) Ways to work:
comp_FOR(.50, .500, .500)  # => FOR = 0.5    = (1 - NPV)
#> [1] 0.5
comp_FOR(.50, .333, .666)  # => FOR = 0.5004 = (1 - NPV)
#> [1] 0.5003751