Compute a probability's (missing) complement and return both.

comp_comp_pair is a function that takes 0, 1, or 2 probabilities (p1 and p2) as inputs. If either of them is missing (NA), it computes the complement of the other one and returns both probabilities.

Usage

comp_comp_pair(p1 = NA, p2 = NA)

Arguments

p1

A numeric probability value (in range from 0 to 1). p1 is optional when p2 is provided.

p2

A numeric probability value (in range from 0 to 1). p2 is optional when p1 is provided.

Value

A vector v containing 2 numeric probability values (in range from 0 to 1): v = c(p1, p2).

Details

comp_comp_pair does nothing when both arguments are provided (i.e., !is.na(p1) & !is.na(p2)) and only issues a warning if both arguments are missing (i.e., is.na(p1) & is.na(p2)).

Inputs are not verified: Use is_prob to verify that an input is a probability and is_complement to verify that two provided values actually are complements.

See also

Examples

# ways to work:
comp_comp_pair(1, 0)   # => 1 0
#> [1] 1 0
comp_comp_pair(0, 1)   # => 0 1
#> [1] 0 1
comp_comp_pair(1, NA)  # => 1 0
#> [1] 1 0
comp_comp_pair(NA, 1)  # => 0 1
#> [1] 0 1

# watch out for:
comp_comp_pair(NA, NA) # => NA NA + warning
#> Warning: One argument (either p1 or p2) is necessary.
#> [1] NA NA
comp_comp_pair(8, 8)   # => 8 8 + NO warning (as is_prob is not verified)
#> [1] 8 8
comp_comp_pair(1, 1)   # => 1 1 + NO warning (as is_complement is not verified)
#> [1] 1 1