Compute overall error rate (err) from probabilities.

comp_err computes overall error rate err from 3 essential probabilities prev, sens, and spec.

Usage

comp_err(prev, sens, spec)

Arguments

prev

The condition's prevalence prev (i.e., the probability of condition being TRUE).

sens

The decision's sensitivity sens (i.e., the conditional probability of a positive decision provided that the condition is TRUE).

spec

The decision's specificity value spec (i.e., the conditional probability of a negative decision provided that the condition is FALSE).

Value

Overall error rate err as a probability (proportion). A warning is provided for NaN values.

Details

comp_err uses comp_acc to compute err as the complement of acc:

err = 1 - acc

See comp_acc and acc for further details and accu for other accuracy metrics and several possible interpretations of accuracy.

See also

Examples

# ways to work:
comp_err(.10, .200, .300)  # => err = 0.71
#> [1] 0.71
comp_err(.50, .333, .666)  # => err = 0.5005
#> [1] 0.5005

# watch out for vectors:
prev.range <- seq(0, 1, by = .1)
comp_err(prev.range, .5, .5)  # => 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5
#>  [1] 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5

# watch out for extreme values:
comp_err(1, 1, 1)  #  => 0
#> [1] 0
comp_err(1, 1, 0)  #  => 0
#> [1] 0

comp_err(1, 0, 1)  #  => 1
#> [1] 1
comp_err(1, 0, 0)  #  => 1
#> [1] 1

comp_err(0, 1, 1)  #  => 0
#> [1] 0
comp_err(0, 1, 0)  #  => 1
#> [1] 1

comp_err(0, 0, 1)  #  => 0
#> [1] 0
comp_err(0, 0, 0)  #  => 1
#> [1] 1