Compute probabilities from (3 essential) probabilities.

comp_prob computes current probability information from 3 essential probabilities (prev, sens or mirt, spec or fart). It returns a list of 13 probabilities prob as its output.

comp_prob(prev = num$prev, sens = num$sens, mirt = NA,
  spec = num$spec, fart = NA, tol = 0.01)

Arguments

prev
The condition's prevalence value prev (i.e., the probability of the condition being TRUE).

sens
The decision's sensitivity value sens (i.e., the conditional probability of a positive decision provided that the condition is TRUE). sens is optional when its complement mirt is provided.

mirt
The decision's miss rate value mirt (i.e., the conditional probability of a negative decision provided that the condition is TRUE). mirt is optional when its complement sens is provided.

spec
The decision's specificity value spec (i.e., the conditional probability of a negative decision provided that the condition is FALSE). spec is optional when its complement fart is provided.

fart
The decision's false alarm rate fart (i.e., the conditional probability of a positive decision provided that the condition is FALSE). fart is optional when its complement spec is provided.

tol
A numeric tolerance value for is_complement. Default: tol = .01.

Value

A list prob containing 13 probability values.

Details

comp_prob assumes that a sufficient and consistent set of essential probabilities (i.e., prev and either sens or its complement mirt, and either spec or its complement fart) is provided.

comp_prob computes and returns a full set of basic and various derived probabilities (e.g., the probability of a positive decision ppod, the probability of a correct decision acc, the predictive values PPV and NPV, as well as their complements FDR and FOR) in its output of a list prob.

Extreme probabilities (sets containing two or more probabilities of 0 or 1) may yield unexpected values (e.g., predictive values PPV or NPV turning NaN when is_extreme_prob_set evaluates to TRUE).

comp_prob is the probability counterpart to the frequency function comp_freq.

Key relationships between probabilities and frequencies:

Three perspectives on a population:

A population of N individuals can be split into 2 subsets of frequencies in 3 different ways:

by condition:

N = cond_true + cond_false

The frequency cond_true depends on the prevalence prev and the frequency cond_false depends on the prevalence's complement 1 - prev.

by decision:

N = dec_pos + dec_neg

The frequency dec_pos depends on the proportion of positive decisions ppod and the frequency dec_neg depends on the proportion of negative decisions 1 - ppod.

by accuracy (i.e., correspondence of decision to condition):

N = dec_cor + dec_err

Each perspective combines 2 pairs of the 4 essential probabilities (hi, mi, fa, cr).

When providing probabilities, the population size N is a free parameter (independent of the essential probabilities prev, sens, and spec).

If N is unknown (NA), a suitable minimum value can be computed by comp_min_N.

Defining probabilities in terms of frequencies:

Probabilities are -- determine, describe, or are defined as -- the relationships between frequencies. Thus, they can be computed as ratios between frequencies:

prevalence prev:

prev = cond_true/N = (hi + mi) / (hi + mi + fa + cr)

sensitivity sens:

sens = hi/cond_true = hi / (hi + mi) = (1 - mirt)

miss rate mirt:

mirt = mi/cond_true = mi / (hi + mi) = (1 - sens)

specificity spec:

spec = cr/cond_false = cr / (fa + cr) = (1 - fart)

false alarm rate fart:

fart = fa/cond_false = fa / (fa + cr) = (1 - spec)

proportion of positive decisions ppod:

ppod = dec_pos/N = (hi + fa) / (hi + mi + fa + cr)

positive predictive value PPV:

PPV = hi/dec_pos = hi / (hi + fa) = (1 - FDR)

negative predictive value NPV:

NPV = cr/dec_neg = cr / (mi + cr) = (1 - FOR)

false detection rate FDR:

FDR = fa/dec_pos = fa / (hi + fa) = (1 - PPV)

false omission rate FOR:

FOR = mi/dec_neg = mi / (mi + cr) = (1 - NPV)

accuracy acc:

acc = dec_cor/N = (hi + cr) / (hi + mi + fa + cr)

rate of hits, given accuracy p_acc_hi:

p_acc_hi = hi/dec_cor = (1 - cr/dec_cor)

rate of false alarms, given inaccuracy p_err_fa:

p_err_fa = fa/dec_err = (1 - mi/dec_err)

Note: When frequencies are rounded (by round = TRUE in comp_freq), probabilities computed from freq may differ from exact probabilities.

Functions translating between representational formats: comp_prob_prob, comp_prob_freq, comp_freq_prob, comp_freq_freq (see documentation of comp_prob_prob for details).

See also

Examples

# Basics:
comp_prob(prev = .11, sens = .88, spec = .77)                        # => ok: PPV = 0.3210614

#> $prev
#> [1] 0.11
#> 
#> $sens
#> [1] 0.88
#> 
#> $mirt
#> [1] 0.12
#> 
#> $spec
#> [1] 0.77
#> 
#> $fart
#> [1] 0.23
#> 
#> $ppod
#> [1] 0.3015
#> 
#> $PPV
#> [1] 0.3210614
#> 
#> $FDR
#> [1] 0.6789386
#> 
#> $NPV
#> [1] 0.9811024
#> 
#> $FOR
#> [1] 0.01889764
#> 
#> $acc
#> [1] 0.7821
#> 
#> $p_acc_hi
#> [1] 0.1237693
#> 
#> $p_err_fa
#> [1] 0.9394218
#> 

comp_prob(prev = .11, sens = NA, mirt = .12, spec = NA, fart = .23)  # => ok: PPV = 0.3210614

#> $prev
#> [1] 0.11
#> 
#> $sens
#> [1] 0.88
#> 
#> $mirt
#> [1] 0.12
#> 
#> $spec
#> [1] 0.77
#> 
#> $fart
#> [1] 0.23
#> 
#> $ppod
#> [1] 0.3015
#> 
#> $PPV
#> [1] 0.3210614
#> 
#> $FDR
#> [1] 0.6789386
#> 
#> $NPV
#> [1] 0.9811024
#> 
#> $FOR
#> [1] 0.01889764
#> 
#> $acc
#> [1] 0.7821
#> 
#> $p_acc_hi
#> [1] 0.1237693
#> 
#> $p_err_fa
#> [1] 0.9394218
#> 

comp_prob()          # => ok, using current defaults

#> $prev
#> [1] 0.25
#> 
#> $sens
#> [1] 0.85
#> 
#> $mirt
#> [1] 0.15
#> 
#> $spec
#> [1] 0.75
#> 
#> $fart
#> [1] 0.25
#> 
#> $ppod
#> [1] 0.4
#> 
#> $PPV
#> [1] 0.53125
#> 
#> $FDR
#> [1] 0.46875
#> 
#> $NPV
#> [1] 0.9375
#> 
#> $FOR
#> [1] 0.0625
#> 
#> $acc
#> [1] 0.775
#> 
#> $p_acc_hi
#> [1] 0.2741935
#> 
#> $p_err_fa
#> [1] 0.8333333
#> 

length(comp_prob())  # => 13 probabilities

#> [1] 13

# Ways to work:
comp_prob(.99, sens = .99, spec = .99)              # => ok: PPV = 0.999898

#> $prev
#> [1] 0.99
#> 
#> $sens
#> [1] 0.99
#> 
#> $mirt
#> [1] 0.01
#> 
#> $spec
#> [1] 0.99
#> 
#> $fart
#> [1] 0.01
#> 
#> $ppod
#> [1] 0.9802
#> 
#> $PPV
#> [1] 0.999898
#> 
#> $FDR
#> [1] 0.00010202
#> 
#> $NPV
#> [1] 0.5
#> 
#> $FOR
#> [1] 0.5
#> 
#> $acc
#> [1] 0.99
#> 
#> $p_acc_hi
#> [1] 0.99
#> 
#> $p_err_fa
#> [1] 0.01
#> 

comp_prob(.99, sens = .90, spec = NA, fart = .10)   # => ok: PPV = 0.9988789

#> $prev
#> [1] 0.99
#> 
#> $sens
#> [1] 0.9
#> 
#> $mirt
#> [1] 0.1
#> 
#> $spec
#> [1] 0.9
#> 
#> $fart
#> [1] 0.1
#> 
#> $ppod
#> [1] 0.892
#> 
#> $PPV
#> [1] 0.9988789
#> 
#> $FDR
#> [1] 0.001121076
#> 
#> $NPV
#> [1] 0.08333333
#> 
#> $FOR
#> [1] 0.9166667
#> 
#> $acc
#> [1] 0.9
#> 
#> $p_acc_hi
#> [1] 0.99
#> 
#> $p_err_fa
#> [1] 0.01
#> 

# Watch out for extreme cases:
comp_prob(1, sens = 0, spec = 1)      # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 1
#> 
#> $fart
#> [1] 0
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

comp_prob(1, sens = 0, spec = 0)      # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 0
#> 
#> $fart
#> [1] 1
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

comp_prob(1, sens = 0, spec = NA, fart = 0)  # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 1
#> 
#> $fart
#> [1] 0
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

comp_prob(1, sens = 0, spec = NA, fart = 1)  # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 0
#> 
#> $fart
#> [1] 1
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

# Watch out for extreme cases:
comp_prob(1, sens = 0, spec = 1)      # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 1
#> 
#> $fart
#> [1] 0
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

comp_prob(1, sens = 0, spec = 0)      # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 0
#> 
#> $fart
#> [1] 1
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

comp_prob(1, sens = 0, spec = NA, fart = 0)  # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 1
#> 
#> $fart
#> [1] 0
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

comp_prob(1, sens = 0, spec = NA, fart = 1)  # => ok, but with warnings (as PPV & FDR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 0):
#>   N mi (FN) cases; 0 cond_false or dec_true cases; PPV = NaN.

#> Warning: PPV is NaN.

#> Warning: FDR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 0
#> 
#> $mirt
#> [1] 1
#> 
#> $spec
#> [1] 0
#> 
#> $fart
#> [1] 1
#> 
#> $ppod
#> [1] 0
#> 
#> $PPV
#> [1] NaN
#> 
#> $FDR
#> [1] NaN
#> 
#> $NPV
#> [1] 0
#> 
#> $FOR
#> [1] 1
#> 
#> $acc
#> [1] 0
#> 
#> $p_acc_hi
#> [1] NaN
#> 
#> $p_err_fa
#> [1] 0
#> 

comp_prob(1, sens = 1, spec = 0)      # => ok, but with warnings (as NPV & FOR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 1):
#>   N hi (TP) cases; 0 cond_false or dec_false cases; NPV = NaN.

#> Warning: NPV is NaN.

#> Warning: FOR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 1
#> 
#> $mirt
#> [1] 0
#> 
#> $spec
#> [1] 0
#> 
#> $fart
#> [1] 1
#> 
#> $ppod
#> [1] 1
#> 
#> $PPV
#> [1] 1
#> 
#> $FDR
#> [1] 0
#> 
#> $NPV
#> [1] NaN
#> 
#> $FOR
#> [1] NaN
#> 
#> $acc
#> [1] 1
#> 
#> $p_acc_hi
#> [1] 1
#> 
#> $p_err_fa
#> [1] NaN
#> 

comp_prob(1, sens = 1, spec = 1)      # => ok, but with warnings (as NPV & FOR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 1):
#>   N hi (TP) cases; 0 cond_false or dec_false cases; NPV = NaN.

#> Warning: NPV is NaN.

#> Warning: FOR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 1
#> 
#> $mirt
#> [1] 0
#> 
#> $spec
#> [1] 1
#> 
#> $fart
#> [1] 0
#> 
#> $ppod
#> [1] 1
#> 
#> $PPV
#> [1] 1
#> 
#> $FDR
#> [1] 0
#> 
#> $NPV
#> [1] NaN
#> 
#> $FOR
#> [1] NaN
#> 
#> $acc
#> [1] 1
#> 
#> $p_acc_hi
#> [1] 1
#> 
#> $p_err_fa
#> [1] NaN
#> 

comp_prob(1, sens = 1, spec = NA, fart = 0)  # => ok, but with warnings (as NPV & FOR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 1):
#>   N hi (TP) cases; 0 cond_false or dec_false cases; NPV = NaN.

#> Warning: NPV is NaN.

#> Warning: FOR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 1
#> 
#> $mirt
#> [1] 0
#> 
#> $spec
#> [1] 1
#> 
#> $fart
#> [1] 0
#> 
#> $ppod
#> [1] 1
#> 
#> $PPV
#> [1] 1
#> 
#> $FDR
#> [1] 0
#> 
#> $NPV
#> [1] NaN
#> 
#> $FOR
#> [1] NaN
#> 
#> $acc
#> [1] 1
#> 
#> $p_acc_hi
#> [1] 1
#> 
#> $p_err_fa
#> [1] NaN
#> 

comp_prob(1, sens = 1, spec = NA, fart = 1)  # => ok, but with warnings (as NPV & FOR are NaN)

#> Warning: Extreme case (prev = 1 & sens = 1):
#>   N hi (TP) cases; 0 cond_false or dec_false cases; NPV = NaN.

#> Warning: NPV is NaN.

#> Warning: FOR is NaN.

#> Warning: Some derived prob values are peculiar. Check for extreme probabilities!

#> $prev
#> [1] 1
#> 
#> $sens
#> [1] 1
#> 
#> $mirt
#> [1] 0
#> 
#> $spec
#> [1] 0
#> 
#> $fart
#> [1] 1
#> 
#> $ppod
#> [1] 1
#> 
#> $PPV
#> [1] 1
#> 
#> $FDR
#> [1] 0
#> 
#> $NPV
#> [1] NaN
#> 
#> $FOR
#> [1] NaN
#> 
#> $acc
#> [1] 1
#> 
#> $p_acc_hi
#> [1] 1
#> 
#> $p_err_fa
#> [1] NaN
#> 

# Ways to fail:
comp_prob(NA, 1, 1, NA)  # => only warning: invalid set (prev not numeric)

#> Warning: Invalid probabilities. Please enter a valid set of essential probabilities.

comp_prob(8,  1, 1, NA)  # => only warning: prev no probability

#> Warning: Invalid probabilities. Please enter a valid set of essential probabilities.

comp_prob(1,  8, 1, NA)  # => only warning: sens no probability

#> Warning: Invalid probabilities. Please enter a valid set of essential probabilities.

comp_prob(1,  1, 1,  1)  # => only warning: is_complement not in tolerated range

#> Warning: Probabilities (p1 and p2) are not complements (in tolerated range).

#> Warning: Invalid probabilities. Please enter a valid set of essential probabilities.