Compute a decision's sensitivity from its miss rate.

comp_sens is a conversion function that takes a miss rate mirt -- given as a probability (i.e., a numeric value in the range from 0 to 1) -- as its input, and returns the corresponding sensitivity sens -- also as a probability -- as its output.

comp_sens(mirt)

Arguments

mirt
The decision's miss rate mirt as a probability.

Value

The decision's sensitivity sens as a probability.

Details

The sensitivity sens and miss rate mirt are complements (sens = (1 - mirt)) and both features of the decision process (e.g., a diagnostic test).

The function comp_sens is complementary to the conversion function comp_mirt and uses the generic function comp_complement.

See also

Examples

comp_sens(2)                       # => NA + warning (beyond range)

#> [1] NA

comp_sens(1/3)                     # => 0.6666667

#> [1] 0.6666667

comp_sens(comp_complement(0.123))  # => 0.123

#> [1] 0.123