Number of individuals for which the condition is false.

cond_false is a frequency that describes the number of individuals in the current population N for which the condition is FALSE (i.e., actually false cases).

Usage

cond_false

Format

An object of class numeric of length 1.

Details

Key relationships:

to probabilities: The frequency of cond_false individuals depends on the population size N and the complement of the condition's prevalence 1 - prev and is split further into two subsets of fa by the false alarm rate fart and cr by the specificity spec.

Perspectives:

by condition:

The frequency cond_false is determined by the population size N times the complement of the prevalence (1 - prev):

cond_false= N x (1 - prev)

by decision:

a. The frequency fa is determined by cond_false times the false alarm rate fart = (1 - spec) (aka. FPR):

fa = cond_false x fart = cond_false x (1 - spec)

b. The frequency cr is determined by cond_false times the specificity spec = (1 - fart):

cr = cond_false x spec = cond_false x (1 - fart)

to other frequencies: In a population of size N the following relationships hold:

N = cond_true + cond_false (by condition)

N = dec_pos + dec_neg (by decision)

N = dec_cor + dec_err (by correspondence of decision to condition)

N = hi + mi + fa + cr (by condition x decision)

Current frequency information is computed by comp_freq and contained in a list freq.

References

Consult Wikipedia: Confusion matrix for additional information.

See also

Examples

cond_false <- 1000 * .90  # => sets cond_false to 90% of 1000 = 900 cases.
is_freq(cond_false)       # => TRUE
#> [1] TRUE
is_prob(cond_false)       # => FALSE, as cond_false is no probability [but (1 - prev) and spec are]
#> [1] FALSE