Number of individuals for which the condition is true.

cond_true is a frequency that describes the number of individuals in the current population N for which the condition is TRUE (i.e., actually true cases).

cond_true

Format

An object of class numeric of length 1.

Details

Key relationships:

to probabilities: The frequency of cond_true individuals depends on the population size N and the condition's prevalence prev and is split further into two subsets of hi by the sensitivity sens and mi by the miss rate mirt.

Perspectives:

by condition:

The frequency cond_true is determined by the population size N times the prevalence prev:

cond_true = N x prev

by decision:

a. The frequency hi is determined by cond_true times the sensitivity sens (aka. hit rate HR):

hi = cond_true x sens

b. The frequency mi is determined by cond_true times the miss rate mirt = (1 - sens):

mi = cond_true x mirt = cond_true x (1 - sens)

to other frequencies: In a population of size N the following relationships hold:

N = cond_true + cond_false (by condition)

N = dec_pos + dec_neg (by decision)

N = dec_cor + dec_err (by correspondence of decision to condition)

N = hi + mi + fa + cr (by condition x decision)

Current frequency information is computed by comp_freq and contained in a list freq.

References

Consult Wikipedia: Confusion matrix for additional information.

See also

Examples

cond_true <- 1000 * .10   # => sets cond_true to 10% of 1000 = 100 cases.
is_freq(cond_true)        # => TRUE
#> [1] TRUE
is_prob(cond_true)        # => FALSE, as cond_true is no probability (but prev and sens are)
#> [1] FALSE