Number of individuals for which the decision is correct.

dec_cor is a frequency that describes the number of individuals in the current population N for which the decision is correct/accurate (i.e., cases in which the decision corresponds to the condition).

dec_cor

Format

An object of class numeric of length 1.

Details

Key relationships:

to probabilities: The frequency of dec_cor individuals depends on the population size N and the accuracy acc.

to other frequencies: In a population of size N the following relationships hold:

N = cond_true + cond_false (by condition)

N = dec_pos + dec_neg (by decision)

N = dec_cor + dec_err (by correspondence of decision to condition)

dec_cor = hi + cr

dec_err = mi + fa

N = hi + mi + fa + cr (by condition x decision)

correspondence: When not rounding the frequencies of freq then

dec_cor = N x acc = hi + cr

(i.e., dec_cor corresponds to the sum of true positives hi and true negatives cr.

Current frequency information is computed by comp_freq and contained in a list freq.

References

Consult Wikipedia: Confusion matrix for additional information.

See also

Examples

dec_cor <- 1000 * .50   # => sets dec_cor to 50% of 1000 = 500 cases.
is_freq(dec_cor)        # => TRUE
#> [1] TRUE
is_prob(dec_cor)        # => FALSE, as dec_cor is no probability (but acc, bacc/wacc ARE)
#> [1] FALSE