Number of individuals for which the decision is erroneous.

dec_err is a frequency that describes the number of individuals in the current population N for which the decision is incorrect or erroneous (i.e., cases in which the decision does not correspond to the condition).

Usage

dec_err

Format

An object of class numeric of length 1.

Details

Key relationships:

to probabilities: The frequency of dec_err individuals depends on the population size N and is equal to the sum of false negatives mi and false positives fa.

to other frequencies: In a population of size N the following relationships hold:

N = cond_true + cond_false (by condition)

N = dec_pos + dec_neg (by decision)

N = dec_cor + dec_err (by correspondence of decision to condition)

dec_cor = hi + cr

dec_err = mi + fa

N = hi + mi + fa + cr (by condition x decision)

Current frequency information is computed by comp_freq and contained in a list freq.

References

Consult Wikipedia: Confusion matrix for additional information.

See also

Examples

dec_err <- 1000 * .50   # => sets dec_err to 50% of 1000 = 500 cases.
is_freq(dec_err)        # => TRUE
#> [1] TRUE
is_prob(dec_err)        # => FALSE, as dec_err is no probability (but acc, bacc/wacc ARE)
#> [1] FALSE