Number of individuals for which the decision is negative.

dec_neg is a frequency that describes the number of individuals in the current population N for which the decision is negative (i.e., cases not called or not predicted).

dec_neg

Format

An object of class numeric of length 1.

Details

Key relationships:

to probabilities: The frequency of dec_neg individuals depends on the population size N and the decision's proportion of negative decisions (1 - ppod) and is split further into two subsets of cr by the negative predictive value NPV and mi by the false omission rate FOR = 1 - NPV.

Perspectives:

by condition:

The frequency dec_neg is determined by the population size N times the proportion of negative decisions (1 - ppod):

dec_neg = N x (1 - ppod)

by decision:

a. The frequency cr is determined by dec_neg times the negative predictive value NPV:

cr = dec_neg x NPV

b. The frequency mi is determined by dec_neg times the false omission rate FOR = (1 - NPV):

mi = dec_neg x FOR = dec_neg x (1 - NPV)

to other frequencies: In a population of size N the following relationships hold:

N = cond_true + cond_false (by condition)

N = dec_pos + dec_neg (by decision)

N = dec_cor + dec_err (by correspondence of decision to condition)

N = hi + mi + fa + cr (by condition x decision)

Current frequency information is computed by comp_freq and contained in a list freq.

References

Consult Wikipedia: Confusion matrix for additional information.

See also

Examples

dec_neg <- 1000 * .67   # => sets dec_neg to 67% of 1000 = 670 cases.
is_freq(dec_neg)        # => TRUE

#> [1] TRUE

is_prob(dec_neg)        # => FALSE, as dec_neg is no probability (but ppod, NPV and FOR are)

#> [1] FALSE