Error rate (err) as the probability of an incorrect decision.

err defines the error rate as the complement of accuracy acc or lack of correspondence of decisions to conditions.

Usage

err

Format

An object of class numeric of length 1.

Details

Definition:

err = (1 - acc)

When freq are not rounded (round = FALSE) then

err = dec_err/N = (mi + fa)/N

err is currently not included in prob, but shown in plots.

See err's complement of accuracy acc for computation and accu for current accuracy metrics and several possible interpretations of accuracy.

See also

Examples

err <- .50     # sets a rate of incorrect decisions of 50%
err <- 50/100  # (dec_err) for 50 out of 100 individuals
is_prob(err)   # TRUE
#> [1] TRUE