Frequency of false alarms or false positives (FP).

fa is the frequency of false alarms or false positives (FP) in a population of N individuals.

fa

Format

An object of class numeric of length 1.

Details

Definition: fa is the frequency of individuals for which Condition = FALSE and Decision = TRUE (positive).

fa is a measure of incorrect classifications (type-I-errors), not an individual case.

Relationships:

to probabilities: The frequency fa depends on the false alarm rate fart (aka. false positive rate, FPR) and is conditional on the prevalence prev.

to other frequencies: In a population of size N the following relationships hold:

N = cond_true + cond_false (by condition)

N = dec_pos + dec_neg (by decision)

N = dec_cor + dec_err (by correspondence of decision to condition)

N = hi + mi + fa + cr (by condition x decision)

See also