The false alarm rate (or false positive rate) of a decision process or diagnostic procedure.

fart defines a decision's false alarm rate (or the rate of false positives): The conditional probability of the decision being positive if the condition is FALSE.

Usage

fart

Format

An object of class numeric of length 1.

Details

Understanding or obtaining the false alarm rate fart:

Definition: fart is the conditional probability for an incorrect positive decision given that the condition is FALSE:

fart = p(decision = positive | condition = FALSE)

or the probability of a false alarm.

Perspective: fart further classifies the subset of cond_false individuals by decision (fart = fa/cond_false).

Alternative names: false positive rate (FPR), rate of type-I errors (alpha), statistical significance level, fallout

Relationships:

a. fart is the complement of the specificity spec:

fart = 1 - spec

b. fart is the opposite conditional probability – but not the complement – of the false discovery rate or false detection rate FDR:

FDR = p(condition = FALSE | decision = positive)

In terms of frequencies, fart is the ratio of fa divided by cond_false (i.e., fa + cr):

fart = fa/cond_false = fa/(fa + cr)

Dependencies: fart is a feature of a decision process or diagnostic procedure and a measure of incorrect decisions (false positives).

However, due to being a conditional probability, the value of fart is not intrinsic to the decision process, but also depends on the condition's prevalence value prev.

References

Consult Wikipedia for additional information.

See also

Examples

fart <- .25     # sets a false alarm rate of 25%
fart <- 25/100  # (decision = positive) for 25 out of 100 people with (condition = FALSE)
is_prob(fart)   # TRUE
#> [1] TRUE