Verify a sufficient set of probability inputs.

is_suff_prob_set is a function that takes 3 to 5 probabilities as inputs and verifies that they are sufficient to compute all derived probabilities and combined frequencies for a population of N individuals.

is_suff_prob_set(prev, sens = NA, mirt = NA, spec = NA, fart = NA)

Arguments

prev

The condition's prevalence prev (i.e., the probability of condition being TRUE).

sens

The decision's sensitivity sens (i.e., the conditional probability of a positive decision provided that the condition is TRUE). sens is optional when its complement mirt is provided.

mirt

The decision's miss rate mirt (i.e., the conditional probability of a negative decision provided that the condition is TRUE). mirt is optional when its complement sens is provided.

spec

The decision's specificity value spec (i.e., the conditional probability of a negative decision provided that the condition is FALSE). spec is optional when its complement fart is provided.

fart

The decision's false alarm rate fart (i.e., the conditional probability of a positive decision provided that the condition is FALSE). fart is optional when its complement spec is provided.

Value

A Boolean value: TRUE if the probabilities provided are sufficient, otherwise FALSE.

Details

While no alternative input option for frequencies is provided, specification of the essential probability prev is always necessary.

However, for 2 other essential probabilities there is a choice:

either sens or mirt is necessary (as both are complements).

either spec or fart is necessary (as both are complements).

is_suff_prob_set does not verify the type, range, or consistency of its arguments. See is_prob and is_complement for this purpose.

See also

Examples

# ways to work:
is_suff_prob_set(prev = 1, sens = 1, spec = 1)  # => TRUE
#> [1] TRUE
is_suff_prob_set(prev = 1, mirt = 1, spec = 1)  # => TRUE
#> [1] TRUE
is_suff_prob_set(prev = 1, sens = 1, fart = 1)  # => TRUE
#> [1] TRUE
is_suff_prob_set(prev = 1, mirt = 1, fart = 1)  # => TRUE
#> [1] TRUE

# watch out for:
is_suff_prob_set(prev = 1, sens = 2, spec = 3)  # => TRUE, but is_prob is FALSE
#> [1] TRUE
is_suff_prob_set(prev = 1, mirt = 2, fart = 4)  # => TRUE, but is_prob is FALSE
#> [1] TRUE
is_suff_prob_set(prev = 1, sens = 2, spec = 3, fart = 4)  # => TRUE, but is_prob is FALSE
#> [1] TRUE

## ways to fail:
# is_suff_prob_set()                    # => FALSE + warning (prev missing)
# is_suff_prob_set(prev = 1)            # => FALSE + warning (sens or mirt missing)
# is_suff_prob_set(prev = 1, sens = 1)  # => FALSE + warning (spec or fart missing)