Frequency of misses or false negatives (FN).

mi is the frequency of misses or false negatives (FN) in a population of N individuals.

Usage

mi

Format

An object of class numeric of length 1.

Details

Definition: mi is the frequency of individuals for which Condition = TRUE and Decision = FALSE (negative).

mi is a measure of incorrect classifications (type-II errors), not an individual case.

Relationships:

to probabilities: The frequency mi depends on the miss rate mirt (aka. false negative rate, FNR) and is conditional on the prevalence prev.

to other frequencies: In a population of size N the following relationships hold:

N = cond_true + cond_false (by condition)

N = dec_pos + dec_neg (by decision)

N = dec_cor + dec_err (by correspondence of decision to condition)

N = hi + mi + fa + cr (by condition x decision)

See also