Frame a 2x2 matrix (from data or description).

mlm_frame creates a 2x2 matrix (i.e., a numeric contingency table of frequency counts) from data or a description of its contents and layout.

mlm_frame(
  data,
  x,
  y,
  z = NA,
  z_val = NA,
  freq_var = NA,
  x_name = NA,
  y_name = NA,
  x_levels = NA,
  y_levels = NA
)

Arguments

data

A data frame (with at least 2 binary variables x and y, denoting individual cases or a contingency table with frequency counts) or a vector with 4 basic frequency counts (abcd, in by-row direction).

x

Variable of data used for x-dimension of 2x2 matrix.

y

Variable of data used for y-dimension of 2x2 matrix.

z

Variable of data used for conditionalizing data.

z_val

Value of z used for conditionalizing data.

freq_var

Variable containing the frequency count of each case when data provides a contingency table (i.e., each row provides a case description).

x_name

Name of x-dimension in 2x2 matrix.

y_name

Name of y-dimension in 2x2 matrix.

x_levels

Order of category levels on x-dimension (values must be present in data).

y_levels

Order of category levels on y-dimension (values must be present in data).

Value

A 2x2 matrix (as a contingency table of integer values).

Details

mlm_frame supports 4 use cases:

From raw data (as data frame): Each row of data denotes an individual case, and the columns contain at least 2 binary variables x and y.

From a contingency table (as data frame): Each row of data denotes a case description, a numeric variable freq_var denotes the frequency count per case, and other columns contain at least 2 binary variables x and y.

From a description: A numeric vector data provides 4 basic frequency counts (i.e., abcd, read in by-row direction), and a specification of the matrix layout (i.e., names of dimensions x and y, and the names/orders of category levels).

From a table (as text): A character string data that can be interpreted by read.ftable of the stats package.

Use cases 1. and 2. allow conditionalizing data on a value z_val of a variable z.

See also

Examples

# 1. From raw data (denoting individual cases):
s <- riskyr(hi = 8, mi = 2, fa = 95, cr = 895)
d <- write_popu(s)
mlm_frame(d, x = "True condition", y = "Outcome")
#> mlm_frame (case 1): Creating mx from raw data (individual cases)
#>           True condition
#> Outcome    present absent
#>   positive       8     95
#>   negative       2    895

# 2. From a contingency table (df) with a variable denoting
#    the frequency counts of each case:
c_tab <- as.data.frame(Titanic)
mlm_frame(c_tab, x = "Sex", y = "Survived", freq_var = "Freq")
#> mlm_frame (case 2): Creating df from aggregated data (contingency table)
#> mlm_frame (case 3): Creating mx from frequency counts (abcd) and layout description
#>         Sex
#> Survived Male Female
#>      No  1364    126
#>      Yes  367    344
mlm_frame(c_tab, x = "Sex", y = "Survived", freq_var = "Freq",
          x_name = "Gender", y_name = "Alive",
          x_levels = c("Female", "Male"), y_levels = c("Yes", "No"))
#> mlm_frame (case 2): Creating df from aggregated data (contingency table)
#> mlm_frame (case 3): Creating mx from frequency counts (abcd) and layout description
#>      Gender
#> Alive Female Male
#>   Yes    344  367
#>   No     126 1364

# 3. From a description of 4 frequency counts and matrix layout:
# a. Basics:
mlm_frame(data = 1:4, x = "Condition", y = "Outcome")
#> mlm_frame (case 3): Creating mx from frequency counts (abcd) and layout description
#>        Condition
#> Outcome c_1 c_2
#>     r_1   1   2
#>     r_2   3   4

# b. The mammography problem:
abcd <- c(8, 95, 2, 895)  # 4 frequencies (Gigerenzer & Hoffrage, 1995)
mlm_frame(data = abcd, x = "Condition", y = "Test",
          x_levels = c("cancer", "no cancer"),
          y_levels = c("positive", "negative"))
#> mlm_frame (case 3): Creating mx from frequency counts (abcd) and layout description
#>           Condition
#> Test       cancer no cancer
#>   positive      8        95
#>   negative      2       895

# c. Titanic passengers:
abcd <- c(344, 367, 126, 1364)
mlm_frame(data = abcd, x = "sex", y = "survived",
          x_levels = c("female", "male"), y_levels = c(1, 0))
#> mlm_frame (case 3): Creating mx from frequency counts (abcd) and layout description
#>         sex
#> survived female male
#>        1    344  367
#>        0    126 1364
mlm_frame(data = abcd, x = "Gender", y = "Alive",
          x_levels = c("Female", "Male"), y_levels = c("Yes", "No"))
#> mlm_frame (case 3): Creating mx from frequency counts (abcd) and layout description
#>      Gender
#> Alive Female Male
#>   Yes    344  367
#>   No     126 1364

# 4. From a table (as text):
tbl_txt <-
"   Condition Cancer no_Cancer
Test
positive      8        95
negative      2       895"
mlm_frame(tbl_txt)
#> mlm_frame (case 4): Creating mx from text data
#>           Condition
#> Test       Cancer no_Cancer
#>   positive      8        95
#>   negative      2       895