Transform a 2x2 matrix (into a table of probabilities/conditional probabilities).

mlm_trans converts a 2x2 matrix mx (i.e., a numeric contingency table of frequency counts) into a table of probabilities or marginal probabilities.

mlm_trans(mx, margin = 0, as_pc = FALSE, n_digits = 3)

Arguments

mx

A 2x2 matrix (as numeric contingency table, required).

margin

Margin to conditionalize table values (numeric, from 0 to 3). Default: margin = 0 (i.e., unconditional probabilities).

as_pc

Boolean: Convert probabilities into percentages? Default: as_pc = FALSE.

n_digits

Number of decimal places to which result is rounded. Default: n_digits = 3.

Value

A numeric 2x2 table.

Details

By default, mlm_trans converts a 2x2 matrix of frequency counts into the corresponding probability values (margin = 0).

Setting margin to 1 to 3 computes conditional probabilities in a by row (margin = 1), by column (margin = 2), or by diagonal (margin = 3) fashion.

The tables for margin = 0 to margin = 2 are computed by the base R function prop.table.

See also

Examples

# The mammography problem:
abcd <- c(8, 95, 2, 895)  # Frequencies (Gigerenzer & Hoffrage, 1995)
mp <- mlm_frame(data = abcd, x = "Condition", y = "Test",
                x_levels = c("cancer", "no cancer"),
                y_levels = c("positive", "negative"))
#> mlm_frame (case 3): Creating mx from frequency counts (abcd) and layout description

mlm_trans(mp)
#>           Condition
#> Test       cancer no cancer
#>   positive  0.008     0.095
#>   negative  0.002     0.895
mlm_trans(mp, margin = 0)  # by-cell: probabilities
#>           Condition
#> Test       cancer no cancer
#>   positive  0.008     0.095
#>   negative  0.002     0.895
mlm_trans(mp, margin = 1)  # by-row: conditional prob
#>           Condition
#> Test       cancer no cancer
#>   positive  0.078     0.922
#>   negative  0.002     0.998
mlm_trans(mp, margin = 2)  # by-col
#>           Condition
#> Test       cancer no cancer
#>   positive  0.800     0.096
#>   negative  0.200     0.904
mlm_trans(mp, margin = 3)  # by-diagonals
#>           Condition
#> Test       cancer no cancer
#>   positive  0.009     0.979
#>   negative  0.021     0.991

mlm_trans(mp, as_pc = TRUE, n_digits = 2)  # as percentages
#>           Condition
#> Test       cancer no cancer
#>   positive    0.8       9.5
#>   negative    0.2      89.5

# The following must sum to 1:
sum(mlm_trans(mp, margin = 0))      # 4 cell values
#> [1] 1
rowSums(mlm_trans(mp, margin = 1))  # 2 row sums
#> positive negative 
#>        1        1 
colSums(mlm_trans(mp, margin = 2))  # 2 col sums
#>    cancer no cancer 
#>         1         1