Alternative color palette for black-and-white (greyscale) graphs.

pal_bw is initialized to a vector of named elements (colors) to define an alternative (black-and-white, b/w) scenario color scheme.

pal_bw

Format

An object of class character of length 16.

Details

Note that pal_bw uses various shades of grey for frequency boxes so that their bounds remain visible on a white background when f_lwd = 0 (as per default for most graphs).

See pal_bwp for a stricter version that enforces black text and lines on white boxes (e.g., for printing purposes).

See pal for default color information.

Assign pal <- pal_bw to use as default color scheme throughout the riskyr package.

See also

Examples

pal_bw        # shows all color names and current values

#>           N   cond_true  cond_false     dec_pos     dec_neg     dec_cor 
#> "#F2F2F2FC" "#E6E6E6FC" "#CCCCCCFC" "#D9D9D9FC" "#B3B3B3FC" "#BFBFBFFC" 
#>     dec_err          hi          mi          fa          cr         ppv 
#> "#999999FC" "#CCCCCCFC" "#999999FC" "#808080FC" "#B3B3B3FC" "#999999FC" 
#>         npv         txt         brd          bg 
#> "#737373FC" "#000000FC" "#1A1A1AFC"     "white" 

pal_bw["hi"]  # shows the current color for hits (true positives, TP)

#>          hi 
#> "#CCCCCCFC" 

pal_bw["hi"] <- "gold" # defines a new color for hits (true positives, TP)