Color palette for cumulative risk curve.

pal_crisk defines a default color palette for the plot_crisk function (as a named vector).

Usage

pal_crisk

Format

An object of class character of length 10.

Details

Color names and referents in plots generated by plot_crisk:

"cum": Cumulative risk curve

"rinc": Relative risk increments

"txt": Text labels

"aux": Auxiliary labels and lines

"high": Highlighting elements

"pas": Past/passed risk

"rem": Remaining risk

"delta": Delta-X- and -Y increments

"poly": Polygon of increments

"popu": Population partitions

See also

Examples

pal_crisk  # show color palette (and names)
#>               cum              rinc               txt               aux 
#>     "deepskyblue"         "#D23430"           "black"          "grey50" 
#>              high               pas               rem             delta 
#>        "deeppink"         "#8E8C84" "lightgoldenrod1"         "#284A6C" 
#>              poly              popu 
#>     "deepskyblue"         "#80B139"