Modern and reduced color palette (in green/blue/bw).

pal_mod is initialized to a vector of named colors to define a reduced modern scenario color scheme (in green/blue/bw).

pal_mbw

Format

An object of class character of length 16.

Details

See pal_org for original color information; pal_mod for a richer modern color palette; and pal_bw for a more reduced black-and-white color palette.

Assign pal <- pal_mbw to use as default color scheme throughout the riskyr package.

See also

Examples

pal_mbw        # shows all color names and current values

#>            N    cond_true   cond_false      dec_pos      dec_neg      dec_cor 
#>  "#E6E6E6FC"  "#CCCCCCFC"  "#BFBFBFFC"  "#D9D9D9FC"  "#A6A6A6FC" "olivedrab4" 
#>      dec_err           hi           mi           fa           cr          ppv 
#> "steelblue3"    "#B8D989"    "#73C8EA"    "#1D95C6"    "#80B139"    "#F26418" 
#>          npv          txt          brd           bg 
#>    "#1D95C6"  "#000000FC"    "#3E3F3A"      "white" 

pal_mbw["hi"]  # shows the current color for hits (true positives, TP)

#>        hi 
#> "#B8D989" 

pal_mbw["hi"] <- "gold" # defines a new color for hits (true positives, TP)