Alternative color palette for unikn.

pal_unikn is initialized to a vector of named elements (colors) to define an alternative (unikn) scenario color scheme.

Usage

pal_unikn

Format

An object of class character of length 16.

Details

See pal for default color information.

Assign pal <- pal_unikn to use as default color scheme throughout the riskyr package.

See also

Examples

pal_unikn        # shows all color names and current values
#>           N   cond_true  cond_false     dec_pos     dec_neg     dec_cor 
#> "#F2F2F2FF"   "#59C7EB"   "#CCEEF9"   "#ECA0B2"   "#8290BB"   "#54BFB7" 
#>     dec_err          hi          mi          fa          cr         ppv 
#>   "#BC7A8F"   "#6AAAB7"   "#D2A6B4"   "#BC7A8F"   "#9CC6CF"   "#E0607E" 
#>         npv         txt         brd          bg 
#>   "#3E5496" "#000000FF" "#1A1A1AFF"     "white" 
pal_unikn["hi"]  # shows the current color for hits (true positives, TP)
#>        hi 
#> "#6AAAB7" 
pal_unikn["hi"] <- "grey" # defines a new color for hits (true positives, TP)