Plot a frequency box object.

plot.box is a utility method that allows to plot low level boxes for riskyr plots.

# S3 method for box
plot(x, cur_freq = freq, lbl_txt = txt, col_pal = pal, ...)

Arguments

x

The box (i.e., an object of class box) to be plotted.

cur_freq

Current frequency information (see freq for details).

lbl_txt

Current text information (see txt for details).

col_pal

Current color palette (see pal for details).

...

Additional (graphical) parameters to be passed to the underlying plotting functions.

Details

plot.riskyr also uses the text settings specified in the "riskyr" object.

See also