The proportion (or baseline) of a positive decision (aka. bias).

ppod defines the proportion (baseline probability or rate) of a decision being positive (but not necessarily accurate/correct).

Usage

ppod

Format

An object of class numeric of length 1.

Details

ppod is also known as bias, though the latter term is also used to describe a systematic tendency to deviate in any — rather than just positive — direction.

Understanding or obtaining the proportion of positive decisions ppod:

Definition: ppod is the (non-conditional) probability:

ppod = p(decision = positive)

or the base rate (or baseline probability) of a decision being positive (but not necessarily accurate/correct).

Perspective: ppod classifies a population of N individuals by decision (ppod = dec_pos/N).

ppod is the "by decision" counterpart to prev (which adopts a "by condition" perspective).

Alternative names: base rate of positive decisions (PR), proportion predicted or diagnosed, rate of decision = positive cases

In terms of frequencies, ppod is the ratio of dec_pos (i.e., hi + fa) divided by N (i.e., hi + mi + fa + cr):

ppod = dec_pos/N = (hi + fa)/(hi + mi + fa + cr)

Dependencies: ppod is a feature of the decision process or diagnostic procedure.

However, the conditional probabilities sens, mirt, spec, fart, PPV, and NPV also depend on the condition's prevalence prev.

References

Consult Wikipedia for additional information.

See also

Examples

ppod <- .50     # sets a rate of positive decisions of 50%
ppod <- 50/100  # (decision = TRUE) for 50 out of 100 individuals
is_prob(ppod)   # TRUE
#> [1] TRUE