The prevalence (baseline probability) of a condition.

prev defines a condition's prevalence value (or baseline probability): The probability of the condition being TRUE.

prev

Format

An object of class numeric of length 1.

Details

Understanding or obtaining the prevalence value prev:

Definition: prev is the (non-conditional) probability:

prev = p(condition = TRUE)

or the base rate (or baseline probability) of the condition's occurrence or truth.

In terms of frequencies, prev is the ratio of cond_true (i.e., hi + mi) divided by N (i.e., hi + mi + fa + cr):

prev = cond_true/N = (hi + mi)/(hi + mi + fa + cr)

Perspective: prev classifies a population of N individuals by condition (prev = cond_true/N).

prev is the "by condition" counterpart to ppod (when adopting a "by decision" perspective) and to acc (when adopting a "by accuracy" perspective).

Alternative names: base rate of condition, proportion affected, rate of condition = TRUE cases.

prev is often distinguished from the incidence rate (i.e., the rate of new cases within a certain time period).

Dependencies: prev is a feature of the population and of the condition, but independent of the decision process or diagnostic procedure.

While the value of prev does not depend on features of the decision process or diagnostic procedure, prev must be taken into account when computing the conditional probabilities sens, mirt, spec, fart, PPV, and NPV (as they depend on prev).

References

Consult Wikipedia for additional information.

See also

Examples

prev <- .10     # sets a prevalence value of 10%
prev <- 10/100  # (condition = TRUE) for 10 out of 100 individuals
is_prob(prev)   # TRUE
#> [1] TRUE