A collection of riskyr scenarios from various sources (as list).

scenarios is a list of scenarios of class riskyr collected from the scientific literature and other sources and to be used by visualization and summary functions.

Usage

scenarios

Format

A list with currently 25 scenarios of class riskyr which are each described by 21 variables.

Details

scenarios currently contains the following scenarios (n1 to n12 in English language, n13 to n25 in German language):

Bowel cancer screening

Cab problem

Hemoccult test

Mammography screening

Mammography (freq)

Mammography (prob)

Mushrooms

Musical town

PSA test (baseline)

PSA test (patients)

Psylicraptis screening

Sepsis

Amniozentese (in German language)

HIV-Test 1

HIV-Test 2

HIV-Test 3

HIV-Test 4

Mammografie 1

Mammografie 2

Mammografie 3

Mammografie 4

Nackenfaltentest (NFT) 1

Nackenfaltentest (NFT) 2

Sigmoidoskopie 1

Sigmoidoskopie 2

Variables describing a scenario:

scen_lbl: Text label for current scenario.

scen_lng: Language of current scenario (en/de).

scen_txt: Description text of current scenario.

popu_lbl: Text label for current population.

cond_lbl: Text label for current condition.

cond_true_lbl: Text label for cond_true cases.

cond_false_lbl: Text label for cond_false cases.

dec_lbl: Text label for current decision.

dec_pos_lbl: Text label for dec_pos cases.

dec_neg_lbl: Text label for dec_neg cases.

hi_lbl: Text label for cases of hits hi.

mi_lbl: Text label for cases of misses mi.

fa_lbl: Text label for cases of false alarms fa.

cr_lbl: Text label for cases of correct rejections cr.

prev: Value of current prevalence prev.

sens: Value of current sensitivity sens.

spec: Value of current specificity spec.

fart: Value of current false alarm rate fart.

N: Current population size N.

scen_src: Source information for current scenario.

scen_apa: Source information in APA format.

Note that names of variables (columns) correspond to a subset of init_txt (to initialize txt) and init_num (to initialize num).

The variables scen_src and scen_apa provide a scenario's source information.

The information of scenarios is also contained in an R data frame df_scenarios (and generated from the corresponding .rda file in /data/).

See also