The sensitivity (or hit rate) of a decision process or diagnostic procedure.

sens defines a decision's sensitivity (or hit rate) value: The conditional probability of the decision being positive if the condition is TRUE.

sens

Format

An object of class numeric of length 1.

Details

Understanding or obtaining the sensitivity sens (or hit rate HR):

Definition: sens is the conditional probability for a (correct) positive decision given that the condition is TRUE:

sens = p(decision = positive | condition = TRUE)

or the probability of correctly detecting true cases (condition = TRUE).

Perspective: sens further classifies the subset of cond_true individuals by decision (sens = hi/cond_true).

Alternative names: true positive rate (TPR), hit rate (HR), probability of detection, power = 1 - beta, recall

Relationships:

a. sens is the complement of the miss rate mirt (aka. false negative rate FNR or the rate of Type-II errors):

sens = (1 - miss rate) = (1 - FNR)

b. sens is the opposite conditional probability -- but not the complement -- of the positive predictive value PPV:

PPV = p(condition = TRUE | decision = positive)

In terms of frequencies, sens is the ratio of hi divided by cond_true (i.e., hi + mi):

sens = hi/cond_true = hi/(hi + mi)

Dependencies: sens is a feature of a decision process or diagnostic procedure and a measure of correct decisions (true positives).

Due to being a conditional probability, the value of sens is not intrinsic to the decision process, but also depends on the condition's prevalence value prev.

References

Consult Wikipedia for additional information.

See also

Examples

sens <- .85     # sets a sensitivity value of 85%
sens <- 85/100  # (decision = positive) for 85 out of 100 people with (condition = TRUE)
is_prob(sens)   # TRUE

#> [1] TRUE