The negative predictive value of a decision process or diagnostic procedure.

NPV defines some decision's negative predictive value (NPV): The conditional probability of the condition being FALSE provided that the decision is negative.

Usage

NPV

Format

An object of class numeric of length 1.

Details

Understanding or obtaining the negative predictive value NPV:

Definition: NPV is the conditional probability for the condition being FALSE given a negative decision:

NPV = p(condition = FALSE | decision = negative)

or the probability of a negative decision being correct.

Perspective: NPV further classifies the subset of dec_neg individuals by condition (NPV = cr/dec_neg = cr/(mi + cr)).

Alternative names: true omission rate

Relationships:

a. NPV is the complement of the false omission rate FOR:

NPV = 1 - FOR

b. NPV is the opposite conditional probability – but not the complement – of the specificity spec:

spec = p(decision = negative | condition = FALSE)

In terms of frequencies, NPV is the ratio of cr divided by dec_neg (i.e., cr + mi):

NPV = cr/dec_neg = cr/(cr + mi)

Dependencies: NPV is a feature of a decision process or diagnostic procedure and – similar to the specificity spec – a measure of correct decisions (negative decisions that are actually FALSE).

However, due to being a conditional probability, the value of NPV is not intrinsic to the decision process, but also depends on the condition's prevalence value prev.

References

Consult Wikipedia for additional information.

See also

Examples

NPV <- .95     # sets a negative predictive value of 95%
NPV <- 95/100  # (condition = FALSE) for 95 out of 100 people with (decision = negative)
is_prob(NPV)   # TRUE
#> [1] TRUE