A collection of riskyr scenarios from various sources (as df).

df_scenarios is an R data frame that contains a collection of scenarios from the scientific literature and other sources.

df_scenarios

Format

A data frame with currently 25 rows (i.e., scenarios) and 21 columns (variables describing each scenario):

See scenarios for a list of scenarios and the variables currently contained in df_scenarios.

Note that names of variables (columns) correspond to a subset of init_txt (to initialize txt) and init_num (to initialize num).

The variables scen_src and scen_apa provide a scenario's source information.

Details

When loading riskyr, all scenarios contained in df_scenarios are converted into a list of riskyr objects scenarios.

See also