List current frequency information.

freq is a list of named numeric variables containing 11 frequencies:

freq

Format

An object of class list of length 11.

Details

the population size N

the number of cases for which cond_true

the number of cases for which cond_false

the number of cases for which dec_pos

the number of cases for which dec_neg

the number of cases for which dec_cor

the number of cases for which dec_err

the number of true positives, or hits hi

the number of false negatives, or misses mi

the number of false positives, or false alarms fa

the number of true negatives, or correct rejections cr

These frequencies are computed from basic parameters (contained in num) and computed by using comp_freq.

The list freq is the frequency counterpart to the list containing probability information prob.

Natural frequencies are always expressed in relation to the current population of size N.

Key relationships between frequencies and probabilities (see documentation of comp_freq or comp_prob for details):

Three perspectives on a population:

by condition / by decision / by accuracy.

Defining probabilities in terms of frequencies:

Probabilities can be computed as ratios between frequencies, but beware of rounding issues.

Functions translating between representational formats: comp_prob_prob, comp_prob_freq, comp_freq_prob, comp_freq_freq (see documentation of comp_prob_prob for details).

Visualizations of current frequency information are provided by plot_prism and plot_icons.

See also

Examples

freq <- comp_freq()  # => initialize freq to default parameters
freq                 # => show current values

#> $N
#> [1] 1000
#> 
#> $cond_true
#> [1] 250
#> 
#> $cond_false
#> [1] 750
#> 
#> $dec_pos
#> [1] 400
#> 
#> $dec_neg
#> [1] 600
#> 
#> $dec_cor
#> [1] 774
#> 
#> $dec_err
#> [1] 226
#> 
#> $hi
#> [1] 212
#> 
#> $mi
#> [1] 38
#> 
#> $fa
#> [1] 188
#> 
#> $cr
#> [1] 562
#> 

length(freq)         # => 11 known frequencies

#> [1] 11

names(freq)          # => show names of known frequencies

#>  [1] "N"          "cond_true"  "cond_false" "dec_pos"    "dec_neg"   
#>  [6] "dec_cor"    "dec_err"    "hi"         "mi"         "fa"        
#> [11] "cr"