The miss rate of a decision process or diagnostic procedure.

mirt defines a decision's miss rate value: The conditional probability of the decision being negative if the condition is TRUE.

Usage

mirt

Format

An object of class numeric of length 1.

Details

Understanding or obtaining the miss rate mirt:

Definition: sens is the conditional probability for an incorrect negative decision given that the condition is TRUE:

mirt = p(decision = negative | condition = TRUE)

or the probability of failing to detect true cases (condition = TRUE).

Perspective: mirt further classifies the subset of cond_true individuals by decision (mirt = mi/cond_true).

Alternative names: false negative rate (FNR), rate of type-II errors (beta)

Relationships:

a. mirt is the complement of the sensitivity sens (aka. hit rate HR):

mirt = (1 - sens) = (1 - HR)

b. mirt is the _opposite_ conditional probability – but not the complement – of the false omission rate FOR:

FOR = p(condition = TRUE | decision = negative)

In terms of frequencies, mirt is the ratio of mi divided by cond_true (i.e., hi + mi):

mirt = mi/cond_true = mi/(hi + mi)

Dependencies: mirt is a feature of a decision process or diagnostic procedure and a measure of incorrect decisions (false negatives).

However, due to being a conditional probability, the value of mirt is not intrinsic to the decision process, but also depends on the condition's prevalence value prev.

References

Consult Wikipedia for additional information.

See also

Examples

mirt <- .15     # => sets a miss rate of 15%
mirt <- 15/100  # => (decision = negative) for 15 out of 100 people with (condition = TRUE)