List current values of scenario color palette.

pal is initialized to a vector of named elements (colors) to define the scenario color scheme that is used throughout the riskyr package.

pal

Format

An object of class character of length 16.

Details

All color information corresponding to the current scenario is stored as named colors in a vector pal. To change a color, assign a new color to an existing element name.

pal currently contains colors with the following names:

N Color representing the population of N cases or individuals.

cond_true Color representing cases of cond_true, for which the current condition is TRUE.

cond_false Color representing cases of in cond_false, for which the current condition is FALSE.

dec_pos Color representing cases of dec_pos, for which the current decision is positive.

dec_neg Color representing cases in dec_neg, for which the current decision is negative.

dec_cor Color representing cases of correct decisions dec_cor, for which the current decision is accurate.

dec_err Color representing cases of erroneous decisions dec_err, for which the current decision is inaccurate.

hi Color representing hits or true positives in hi (i.e., correct cases for which the current condition is TRUE and the decision is positive).

mi Color representing misses or false negatives in mi (i.e., incorrect cases for which the current condition is TRUE but the decision is negative).

fa Color representing false alarms or false positives in fa (i.e., incorrect cases for which the current condition is FALSE but the decision is positive).

cr Color representing correct rejections or true negatives in cr (i.e., correct cases for which the current condition is FALSE and the decision is negative).

ppv Color representing positive predictive values PPV (i.e., the conditional probability that the condition is TRUE, provided that the decision is positive).

npv Color representing negative predictive values NPV (i.e., the conditional probability that the condition is FALSE, provided that the decision is negative).

txt Color used for text labels.

brd Color used for borders.

bg Color used for plot background (used to set par(bg = bg_col)).

Note that color names for frequencies correspond to frequency names, but are different for probabilities (which are written in lowercase and only PPV and NPV have assigned colors).

See also

Examples

pal        # shows all color names and current values

#>                 N         cond_true        cond_false           dec_pos 
#>       "#E6E6E6FC" "lightgoldenrod1"       "#BFBFBFFC"         "#F7A97F" 
#>           dec_neg           dec_cor           dec_err                hi 
#>       "#A6A6A6FC"      "olivedrab4"      "steelblue3"         "#B8D989" 
#>                mi                fa                cr               ppv 
#>         "#73C8EA"         "#1D95C6"         "#80B139"         "#F26418" 
#>               npv               txt               brd                bg 
#>         "#1D95C6"       "#000000FC"         "#3E3F3A"           "white" 

pal["hi"]  # shows the current color for hits (true positives, TP)

#>        hi 
#> "#B8D989" 

pal["hi"] <- "gold"  # defines a new color for hits (true positives, TP)