A table of cases (in the current population).

popu is an R data frame that is computed by comp_popu from the current frequency information (contained in freq). Each individual is represented as a row; columns represent the individual's condition (TRUE or FALSE), a corresponding decision (also encoded as TRUE = positive or FALSE = negative), and its classification (i.e., its case or cell combination, in SDT terms), as true positive (hit hi), false negative (miss mi), false positive (false alarm fa), or true negative (correct rejection cr).

popu

Format

An object of class NULL of length 0.

Value

A data frame popu containing N rows (individual cases) and 3 columns ("Truth", "Decision", "SDT") encoded as ordered factors (with 2, 2, and 4 levels, respectively).

Details

#' popu is initialized to NULL and needs to be computed by calling comp_popu with current parameter settings.

By default, comp_popu uses the current information contained in txt to define text labels.

A visualization of the current population popu is provided by plot_icons.

See also

Examples

popu <- comp_popu()  # => initializes popu with current values of freq and txt
dim(popu)            # => N x 3

#> [1] 1000    3

head(popu)           # => shows head of data frame

#>   Truth     Test Cases
#> 1  True Positive    TP
#> 2  True Positive    TP
#> 3  True Positive    TP
#> 4  True Positive    TP
#> 5  True Positive    TP
#> 6  True Positive    TP