The specificity of a decision process or diagnostic procedure.

spec defines a decision's specificity value (or correct rejection rate): The conditional probability of the decision being negative if the condition is FALSE.

Usage

spec

Format

An object of class numeric of length 1.

Details

Understanding or obtaining the specificity value spec:

Definition: spec is the conditional probability for a (correct) negative decision given that the condition is FALSE:

spec = p(decision = negative | condition = FALSE)

or the probability of correctly detecting false cases (condition = FALSE).

Perspective: spec further classifies the subset of cond_false individuals by decision (spec = cr/cond_false).

Alternative names: true negative rate (TNR), correct rejection rate, 1 - alpha

Relationships:

a. spec is the complement of the false alarm rate fart:

spec = 1 - fart

b. spec is the opposite conditional probability – but not the complement – of the negative predictive value NPV:

NPV = p(condition = FALSE | decision = negative)

In terms of frequencies, spec is the ratio of cr divided by cond_false (i.e., fa + cr):

spec = cr/cond_false = cr/(fa + cr)

Dependencies: spec is a feature of a decision process or diagnostic procedure and a measure of correct decisions (true negatives).

However, due to being a conditional probability, the value of spec is not intrinsic to the decision process, but also depends on the condition's prevalence value prev.

References

Consult Wikipedia for additional information.

See also

Examples

spec <- .75     # sets a specificity value of 75%
spec <- 75/100  # (decision = negative) for 75 out of 100 people with (condition = FALSE)
is_prob(spec)   # TRUE
#> [1] TRUE