Basic text elements.

txt is initialized to a list of named elements to define basic scenario titles and labels.

txt

Format

An object of class list of length 21.

Details

All textual elements that specify generic labels and titles of riskyr scenarios are stored as named elements (of type character) in a list txt. To change an element, assign a new character object to an existing name.

The list txt is used throughout the riskyr package unless a scenario defines scenario-specific text labels (when using the riskyr function).

Note:

Basic text information and some numeric parameters (see num and init_num) are integral parts of a riskyr scenario.

By contrast, basic color information (see pal and init_pal) is not an integral part, but independently defined.

The names of probabilities (see prob) are currently not an integral part of txt and riskyr scenarios (but defined in prob_lbl_def and label_prob).

txt currently contains the following text labels:

scen_lbl The current scenario title (sometimes in Title Caps).

scen_txt A longer text description of the current scenario (which may extend over several lines).

scen_src The source information for the current scenario.

scen_apa The source information in APA format.

scen_lng The language of the current scenario (as character code). Options: "en": English, "de": German.

popu_lbl A general name describing the current population.

N_lbl A short label for the current population popu or sample.

cond_lbl A general name for the condition dimension, or the feature (e.g., some disease) currently considered.

cond_true_lbl A short label for the presence of the current condition or cond_true cases (the condition's true state of being TRUE).

cond_false_lbl A short label for the absence of the current condition or cond_false cases (the condition's true state of being FALSE).

dec_lbl A general name for the decision dimension, or the judgment (e.g., some diagnostic test) currently made.

dec_pos_lbl A short label for positive decisions or dec_pos cases (e.g., predicting the presence of the condition).

dec_neg_lbl A short label for negative decisions or dec_neg cases (e.g., predicting the absence of the condition).

acc_lbl A general name for the accuracy dimension, or the correspondence between the condition currently considered and the decision judgment currently made.

dec_cor_lbl A short label for correct and accurate decisions or dec_cor cases (accurate predictions).

dec_err_lbl A short label for incorrect decisions or dec_err cases (erroneous predictions).

sdt_lbl A general name for all 4 cases/categories/cells of the 2x2 contingency table (e.g., condition x decision, using SDT).

hi_lbl A short label for hits or true positives hi/TP cases (i.e., correct decisions of the presence of the condition, when the condition is actually present).

mi_lbl A short label for misses or false negatives mi/FN cases (i.e., incorrect decisions of the absence of the condition when the condition is actually present).

fa_lbl A short label for false alarms or false positives fa/FP cases (i.e., incorrect decisions of the presence of the condition when the condition is actually absent).

cr_lbl A short label for correct rejections or true negatives cr/TN cases (i.e., a correct decision of the absence of the condition, when the condition is actually absent).

See also

Examples

txt           # Show  all current names and elements
#> $scen_lbl
#> [1] "Scenario"
#> 
#> $scen_txt
#> [1] "Description of the current scenario in a short paragraph of text. This description may include several sentences."
#> 
#> $scen_src
#> [1] "Source information"
#> 
#> $scen_apa
#> [1] "Source information in APA format"
#> 
#> $scen_lng
#> [1] "en/de"
#> 
#> $popu_lbl
#> [1] "Population"
#> 
#> $N_lbl
#> [1] "N"
#> 
#> $cond_lbl
#> [1] "True condition"
#> 
#> $cond_true_lbl
#> [1] "present"
#> 
#> $cond_false_lbl
#> [1] "absent"
#> 
#> $dec_lbl
#> [1] "Outcome"
#> 
#> $dec_pos_lbl
#> [1] "positive"
#> 
#> $dec_neg_lbl
#> [1] "negative"
#> 
#> $acc_lbl
#> [1] "Accuracy"
#> 
#> $dec_cor_lbl
#> [1] "correct"
#> 
#> $dec_err_lbl
#> [1] "incorrect"
#> 
#> $sdt_lbl
#> [1] "Cases"
#> 
#> $hi_lbl
#> [1] "TP"
#> 
#> $mi_lbl
#> [1] "FN"
#> 
#> $fa_lbl
#> [1] "FP"
#> 
#> $cr_lbl
#> [1] "TN"
#> 
txt$scen_lbl  # Show the current scenario label (e.g., used in plot titles)
#> [1] "Scenario"
txt$scen_lbl <- "My example"  # Set a new scenario title