Alternative text labels (TP, FN, FP, TN).

txt_TF is initialized to alternative text labels to define a frequency naming scheme in which (hi, mi, fa, cr) are called (TP, FN, FP, TN).

txt_TF

Format

An object of class list of length 21.

Details

See txt for details and default text information.

Assign txt <- txt_TF to use as default text labels.

See also

Examples

txt_TF        # shows text labels of txt_TF

#> $scen_lbl
#> [1] "Scenario"
#> 
#> $scen_txt
#> [1] "Description of the current scenario in a short paragraph of text. This description may include several sentences."
#> 
#> $scen_src
#> [1] "Source information"
#> 
#> $scen_apa
#> [1] "Source information in APA format"
#> 
#> $scen_lng
#> [1] "en/de"
#> 
#> $popu_lbl
#> [1] "Population"
#> 
#> $N_lbl
#> [1] "N"
#> 
#> $cond_lbl
#> [1] "Truth"
#> 
#> $cond_true_lbl
#> [1] "True"
#> 
#> $cond_false_lbl
#> [1] "False"
#> 
#> $dec_lbl
#> [1] "Test"
#> 
#> $dec_pos_lbl
#> [1] "Positive"
#> 
#> $dec_neg_lbl
#> [1] "Negative"
#> 
#> $acc_lbl
#> [1] "Accuracy"
#> 
#> $dec_cor_lbl
#> [1] "Correct"
#> 
#> $dec_err_lbl
#> [1] "Incorrect"
#> 
#> $sdt_lbl
#> [1] "Cases"
#> 
#> $hi_lbl
#> [1] "TP"
#> 
#> $mi_lbl
#> [1] "FN"
#> 
#> $fa_lbl
#> [1] "FP"
#> 
#> $cr_lbl
#> [1] "TN"
#> 

txt_TF["hi"]  # shows the current label for hits ("TP")

#> $<NA>
#> NULL
#> 

txt_TF["hi"] <- "hit" # defines a new label for hits (true positives, TP)